feat(dotnet): port search functions (Chunk 9)
Mirrors src/semble/search.py:

  - SearchSemantic, SearchBm25, SearchHybrid
  - SortTopK (== np.argsort(-x)[:top_k] for the no-tie case)
  - RrfScores private helper (k=60, matching _RRF_K)
  - SearchHybrid alpha-blends RRF-normalised scores, applies
    BoostMultiChunkFiles + ApplyQueryBoost + RerankTopK

Notes:
  - SearchResult.Score widened to double to match Python's float64
    arithmetic during RRF/boost/penalty composition.
  - Fixtures.MockEncoder mirrors the conftest.py mock_model fixture:
    deterministic seeded unit-norm random vectors via Box-Muller.
    Bit-different from numpy's PCG64 but structurally equivalent for
    the structural assertions the tests rely on (length, score range,
    source label, dedup behavior).

Tests port all 13 cases from tests/test_search.py with the mock-based
load_model parametrisation re-shaped around the .NET IEncoder seam.

+++ dotnet/src/Semble/Types.cs
@@ -57,9 +57,10 @@ public sealed record Chunk(
 }
 
 /// <summary>
-/// A single search result with score and source retriever.
+/// A single search result with score and source retriever. Score is double-precision
+/// to match Python's float64 — matters for hybrid/RRF arithmetic.
 /// </summary>
-public sealed record SearchResult(Chunk Chunk, float Score, SearchMode Source);
+public sealed record SearchResult(Chunk Chunk, double Score, SearchMode Source);
 
 /// <summary>
 /// Statistics about the current index state.
